web-app/sivarama-ayyar/requirements.txt
huggingface_hub==0.25.2
requests
gradio
google-generativeai
bs4